Update amazonannoyances.txt
There is a new section that I saw on the homepage for the first time the other day with livestreamed shopping, or something to that effect. I would definitely consider it an annoyance that should be blocked. It can also be blocked with: ""amazon.*##.amazonlive-widget-padding"" but ""amazon.*###live-flagship-root"" is the div while the live-widget-padding is the class. I also added in this a network rule ""||live-video.net^"" to just block loading of video from live-video.net so their shopping livestreams won't load.